Backend/requirements.txt
asgiref==3.8.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.1
graphviz==0.20.3
greenlet==3.2.2
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.5.0
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scikit-posthocs==0.11.4
scipy==1.15.3
seaborn==0.13.2
shap==0.47.2
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
SQLAlchemy==2.0.41
sqlparse==0.5.3
statsmodels==0.14.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
